Iceshelf devj

+++ fvtools/grid/fvcom_grd.py
@@ -2215,7 +2215,8 @@ class NestROMS2FVCOM:
         if self.oend1 == 1:
             new_x_obc, new_y_obc = [], []
             for n in range(self.x_obc.shape[0]):
-                dist       = np.sqrt((self.x_obc[n] - self.x_obc[n][0])**2 + (self.y_obc[n] - self.y_obc[n][0])**2)
+                # dist       = np.sqrt((self.x_obc[n] - self.x_obc[n][0])**2 + (self.y_obc[n] - self.y_obc[n][0])**2)
+                dist       = np.sqrt(((self.x_obc[n] - self.x_obc[n][0])**2 + (self.y_obc[n] - self.y_obc[n][0])**2).astype(float))
                 i          = np.where(dist>self.R)
                 new_x_obc.append(self.x_obc[n][i])
                 new_y_obc.append(self.y_obc[n][i])
@@ -2224,7 +2225,8 @@ class NestROMS2FVCOM:
         if self.oend2 == 1:
             new_x_obc, new_y_obc = [], []
             for n in range(self.x_obc.shape[0]):
-                dist = np.sqrt((self.x_obc[n] - self.x_obc[n][-1])**2 + (self.y_obc[n] - self.y_obc[n][-1])**2)
+                #dist = np.sqrt((self.x_obc[n] - self.x_obc[n][-1])**2 + (self.y_obc[n] - self.y_obc[n][-1])**2)
+                dist = np.sqrt(((self.x_obc[n] - self.x_obc[n][-1])**2 + (self.y_obc[n] - self.y_obc[n][-1])**2).astype(float))
                 i    = np.where(dist>self.R)
                 new_x_obc.append(self.x_obc[n][i])
                 new_y_obc.append(self.y_obc[n][i])
